pyproject.toml
[tool.poetry]
name = "serviceteam"
version = "0.1.0"
description = "ServiceTEam using crewAI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
langchain = ">=0.1.10,<0.2.0"  # Updated version constraint
PyPDF2 = "3.0.1"
google-generativeai="0.5.4"
python-dotenv="1.0.1"
crewai = { extras = ["tools"], version = "^0.30.11" }
streamlit = "^1.35.0"

[tool.poetry.scripts]
serviceteam = "serviceteam.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"